% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotter_function.R
\name{rez.comp}
\alias{rez.comp}
\title{A Residual Plot Comparison Function}
\usage{
rez.comp(x, y, z)
}
\arguments{
\item{x}{a vector of data}

\item{y}{a vector of numerical data}

\item{z}{a dataframe containing x and y}
}
\description{
This function allows you to compare normal, logged and square rooted residual plots of your data
}
\examples{
rez.comp(cyl,mpg,mtcars)
}
\keyword{residuals}

